Removed string comparisons and unnecesary reference to D1 manifest
